job_container/requirements.txt
aioflask
gunicorn
pandas
numpy
requests
nest-asyncio
pyppeteer
pyppdf
absl-py
google-cloud-storage
google-cloud-logging
google-cloud-spanner
cloud-sql-python-connector
pg8000
sqlalchemy
bs4
google-cloud-bigquery
geojson
google-cloud-tasks